Running art/project.py

+++ art/project.py
@@ -2,6 +2,6 @@ from random import random
 r=16
 c=16
 HighestNoMines = 10
-Nc=random()*HighestNoMines
+Nc=int(random()*HighestNoMines)
 
 print Nc
